requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
openai>=1.50.0
pdfplumber>=0.11.9
pymupdf>=1.26.7
reportlab>=4.4.9
python-dotenv>=1.0.0
